requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
sqlalchemy
numpy
matplotlib
python-dotenv
google-generativeai
pdfplumber
sendgrid
pymupdf
